requirements.txt
Django==3.2.11
django-allauth
django-cors-headers
django-filter
django-heroku
django-nested-admin
django-redis
django-rest-auth
djangorestframework
coreapi
drf-yasg
gunicorn
Pillow
psycopg2-binary
python-decouple
Faker
whitenoise
django-cloudinary-storage

channels_redis
channels
redis
daphne